requirements.txt
line-bot-sdk==3.7
flask==3.0.0
openai>=1.0.0